finish bfs large input

+++ rodinia_3.1/cuda/bfs/kernel2.cu
@@ -19,9 +19,9 @@ The CUDA Kernel for Applying BFS on a loaded Graph. Created By Pawan Harish
 #define _KERNEL2_H_
 
 __global__ void
-Kernel2( bool* g_graph_mask, bool *g_updating_graph_mask, bool* g_graph_visited, bool *g_over, int no_of_nodes)
+Kernel2( bool* g_graph_mask, bool *g_updating_graph_mask, bool* g_graph_visited, bool *g_over, unsigned long long no_of_nodes)
 {
-	int tid = blockIdx.x*MAX_THREADS_PER_BLOCK + threadIdx.x;
+	unsigned long long tid = blockIdx.x*MAX_THREADS_PER_BLOCK + threadIdx.x;
 	if( tid<no_of_nodes && g_updating_graph_mask[tid])
 	{
